Move utility methods from `AbstractEntity` to `Entities`.
